Added bitmap demo
Converted the CLOUDS.raw file to C source code (using HxD hex editor) and figured out what the actual width and height were. The original demo used the wrong numbers so the bitmap was rendered wrong with weird interlacing effects. I derived the actual width and height from observation and experimentation. Looks like the bitmap is a lot higher than needed, and has a big area of nothing in it. I may make it smaller at a later time, if I can ever figure out how to load it into GIMP or something.

I pretty much completely rewrote the demo code to generate the scrolling clouds. The old code loaded the bitmap from SD card which I may add as an option in the future. The current code copies the bitmap from RAM. I should store the data in PROGMEM and copy it block by block.

Also fixed a couple of minor bugs in the StEVE code and added a RegWriteBuffer of course.